requirement.txt
streamlit
langchain-ollama
ollama
Pillow
pytesseract
python-magic
diffusers
transformers
accelerate
dotenv
torch
xformers
wheel
google.oauth2
googleapiclient.discovery 
chromadb 
sentence-transformers
PyPDF2
setuptools